+++ examples/faker/src/gen/models/Item.ts
@@ -1,0 +1,13 @@
+export type Item = {
+  /**
+   * @minLength 3
+   * @maxLength 25
+   * @type string
+   */
+  name: string
+  /**
+   * @minLength 5
+   * @type number
+   */
+  price: number
+}